Fix allocate and initialize bug for diag arrays
A bug was found that was causing a segfault.  If the user specified
that sstflag=n or genflag=n, then there are certain arrays that never
get allocated, however I still had statements in the main line of
code that would try to initialize those arrays to missing values.  I
changed the code so that the initialization statements get included
inside the IF statements for the checks of sstflag and genflag where
the allocate statements take place.