Fix allocatable arrays and add flexible unit nums
A user, Alan Brammer, notified us of a bug in which arrays that are
allocatable -- but not allocated -- are attempted to be accessed,
leading to a seg fault.  This was for arrays such as sst if sstflag was
not set to y/Y, and also for arrays such as q850, rh, spfh, temperature
and omega if genflag was not set to y/Y.  Also, in tave, vint and in the
gettrk executables, the open statements were made more flexible to
accomodate the NCEP method of opening files.  This change was coded
by Hana Jafary from EMC.